pyproject.toml
[project]
name = "wait"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.3",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
